Changed Trakt plugin into a Windows Plugin. Added some GUI classes to help build up some windows. You should delete the process plugin and copy new one to plugins\windows folder.
Note: the main window is currently a launching pad to access different areas of trakt (basically a few GUI buttons), these are not hooked up to anything just yet. We can also make available some content to skinners on the main window so can be just more than a bunch of buttons accessing different areas.