Improved Syncing of Movie library for movies that fail to be matched on Trakt (ie skipped movies). These are not persisted to disk and filtered out on library sync, if the list is older than 7 days then they are re-added to sync list.
Debug log will display what movies need attention in an easy to read way. It will also inform user that the movie info ie either bad or movie needs to be added to themoviedb.org.

Note: Added AlreadyExistsMovies in preparation for API update which returns this in response, once that is done issue #34 can be signed off.