Fixed Fanart not being shown in Shouts window, if launched from Activity Dashboard and local fanart doesnt exist. We now check if the fanart has already been downloaded, if it hasn't then it will go fetch it during shouts retrieval.
It's possible the user had not downloaded the fanart from other GUI windows such as Trending so it makes sense to add this. I have noticed I didn't have fanart on a few occasions.

Also refactored TraktUser DataStructure and it's usage, Shouts and Dashboard will use the simple datastructure as it doesn't get all the other properties such as user stats.