Added filtering support to Trending GUI Windows, can filter watched, watchlisted, rated and collected.
Note: Collected will not work on tvshows yet as its not available as a field in the API. Will add caching of tvdb ids when retrieving library during sync so can compare against that. Also 'Watched' for tvshows only counts if 1 episode is watched not all episodes. Updated skins to follow.