Added ability to 'Search All' people for a selected movie, show, episode in External Trakt SearchBy menu.
This will perform a lookup for all people and display results rather than needing to do one search at a time. Single search is still good as that will return many results for a single person which is good when scrapers don't get the name right.

When doing 'Search All' it only retreives a single result for each person which is the best match.