Added Sync of Custom Lists to MovingPictures from trakt.
Big re-factor of Categories and Filter menu handling in MovingPictures.
- Now using IMDb field to create criteria in node filters. Don't need to worry about the movie being in collection as you do with WhiteList/BlackList filters.
- No longer create Blacklist of all movies in collection when nothing is added to a node, create a dummy criteria filter instead.

Known Issue: When adding a criteria to an empty node, if the node is no longer empty it is not shown in GUI until the MovingPictures browser refreshes. Similarly if you remove criteria such that the filtered movie list is empty it does not remove the node from being visible in the GUI.