Added support for icon overlays on TV Season posters.
Currently will only do Collected and Watched, Ratings and Watchlisted for seasons are not retrieved during TVSeries sync so we do not know yet whether a season is rated or watchlisted.

The Season icons for collected/watched will not take into account the aired_episodes only, the season summary method only returns total count so can only use that to determine if a season is collected/watched.